WIP relocate component, add logic to show or hide mobile menu
